+++ R/residSVD.R
@@ -1,4 +1,4 @@
-#  Copyright (C) 2005, 2012 David Firth and Heather Turner
+#  Copyright (C) 2005, 2012, 2023 David Firth and Heather Turner
 #
 #  This program is free software; you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
@@ -19,14 +19,14 @@ residSVD <- function(model, fac1, fac2, d = 1) {
         fac1 <- eval(match.call()$fac1, Data, parent.frame())
         fac2 <- eval(match.call()$fac2, Data, parent.frame())
     }
-    if (!inherits(model, "glm") && class(model) != "lm") stop(
-                                        "model not of class lm, glm or gnm")
+    if (!inherits(model, "lm")) 
+        stop("model not of a class that inherits from class \"lm\"")
     if (!is.factor(fac1)) stop("fac1 must be a factor")
     if (!is.factor(fac2)) stop("fac2 must be a factor")
     Data <- data.frame(fac1, fac2)
-    if (!is.null(model$na.action)) Data <- Data[-model$na.action, ]
-    weights <- if (!is.null(model$weights)) as.vector(model$weights) else 1
-    X <- data.frame(rw = as.vector(model$residuals) * weights, w = weights)
+    if (!is.null(na.action(model))) Data <- Data[-na.action(model), ]
+    weights <- if (!is.null(weights(model))) as.vector(weights(model)) else 1
+    X <- data.frame(rw = as.vector(residuals(model)) * weights, w = weights)
     X <- lapply(X, tapply, Data, sum, simplify = TRUE)
     X <- X$rw/X$w
     X <- svd(naToZero(X), d, d)